embedding-simple-demo1/requirements.txt
numpy>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
torch>=2.0.0
matplotlib>=3.7.0